requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selene==2.0.0b5
python-dotenv==0.20.0
CaptchaSolver
selenium-chromedriver==1.3